// frontend/src/types/admin.ts
export interface AdminCityCreateRequest {
  name: string;
  state: number;
  slug?: string;
  photo?: File;
  latitude?: number;
  longitude?: number;
  is_major?: boolean;
  is_active?: boolean;
}

export interface AdminCityUpdateRequest extends Partial<AdminCityCreateRequest> {
  is_active?: boolean;
}